% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronavirus.R
\docType{data}
\name{coronavirus}
\alias{coronavirus}
\title{coronavirus}
\format{
A data frame with 9 variables.
}
\source{
Johns Hopkins University Center for Systems Science and Engineering
(JHU CCSE) Coronavirus \href{https://systems.jhu.edu/research/public-health/ncov/}{website}.
}
\usage{
coronavirus
}
\description{
coronavirus
}
\details{
comment:
this dataset have 200960 observe and 9 varibales,which is data,location, data type, latitude, and longitude
This coronavirus dataset has the following fields:
date - The date of the summary
province - The province or state, when applicable
country - The country or region name
lat - Latitude point
long - Longitude point
cases - the number of daily cases (corresponding to the case type)
(find this dataset in the package coronavirus)

The dataset contains the daily summary of Coronavirus cases
(confirmed, death, and recovered), by state/province.
}
\keyword{COVID19}
\keyword{coronavirus}
\keyword{datasets}
