% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_type.R
\name{count_type}
\alias{count_type}
\title{count cases type}
\usage{
count_type(date, data_type)
}
\arguments{
\item{date}{A string, The date at which the case numbers are required}

\item{data_type}{A string, The type of case numbers required - either "Confirmed", "Deaths" or "Recovered"}
}
\value{
An integer, sum of covid cases of required type and date
}
\description{
Get case count at one date
}
\examples{
\dontrun{
get_case_count(data,
  date = "2020-10-01", case_type = "cases_new")
  
count_type( date = "4/4/20", case_type = "cases_new")
}
}
