% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_select.R
\name{countryInput}
\alias{countryInput}
\title{Country Select Input in table}
\usage{
countryInput(x)
}
\arguments{
\item{x}{A country names}
}
\description{
Country Select Input in table
}
\examples{
\dontrun{
countryInput( x = coronavirus$location)
}





}
\author{
chenjie gong
}
