% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Function to launch the shiny application}
\usage{
launch_app()
}
\value{
shinyapp
}
\description{
This function can be used to launch shiny app in package.
}
\examples{
\dontrun{
launch_app()
}

}
\author{
chenjie
}
