% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_countries.R
\name{n_countries}
\alias{n_countries}
\title{n_countries}
\usage{
n_countries(data)
}
\arguments{
\item{data}{a dataset containing a \code{country} column}
}
\value{
sum of country during this epdemic

sum of country during this epdemic
}
\description{
it help people count the country
}
\details{
count the number of countried in dataset
}
\examples{
n_countries(coronavirus)

}
