% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cases.R
\name{plot_cases}
\alias{plot_cases}
\title{plot top 10 cases new contries}
\usage{
plot_cases(data, p)
}
\arguments{
\item{data}{the coronavirus data}

\item{p}{the top10 bar plot}
}
\value{

}
\description{
plot top 10 cases new contries
}
