% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_select.R
\name{selectdate}
\alias{selectdate}
\title{date setup}
\usage{
selectdate(inputId, label = "select one date", x)
}
\arguments{
\item{inputId}{inputid}

\item{label}{label}

\item{x}{x function}
}
\value{

}
\description{
date setup
}
