% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_select.R
\name{type_select}
\alias{type_select}
\title{Select type}
\usage{
type_select(data, choices)
}
\arguments{
\item{data}{the dateset}

\item{choices}{cases_new,deaths_new,recovered_new}
}
\value{
shiny
}
\description{
This function builded to help people select the different type
}
\author{
chenjie gong
}
