% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{launch_app}
\usage{
launch_app()
}
\value{
shinyapp
}
\description{
Function to launch the shiny application
}
\examples{
\dontrun{
launch_app()
}

}
\author{
chenjie
}
