% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_countries.R
\name{n_countries}
\alias{n_countries}
\title{count the number of countried in dataset}
\usage{
n_countries(data)
}
\arguments{
\item{data}{a dataset containing a \code{country} column}
}
\value{

}
\description{
count the number of countried in dataset
}
